temporary remove some user

+++ src/admin/company/company.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, UseGuards, Get, Req, Res, Param, NotFoundException, Post, Body, Patch, Delete } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
-import { ApiBearerAuth, ApiOperation, ApiImplicitQuery } from '@nestjs/swagger';
+import { ApiBearerAuth, ApiOperation, ApiImplicitQuery, ApiImplicitParam } from '@nestjs/swagger';
 import { CompanyService } from './company.service';
 import { CommonFunctionService } from 'src/common/helper/common-function.services';
 import { UpdateCompanyDTO } from './dto/update-company.dto';
@@ -25,7 +25,7 @@ export class CompanyController {
 	constructor(private readonly companyService: CompanyService, private readonly commonFunctionService: CommonFunctionService) { }
 
 	/**
-	 *  function find alll company
+	 *  function find all company
 	 *
 	 * @param {*} req
 	 * @param {*} res
@@ -34,10 +34,6 @@ export class CompanyController {
 	@Get()
 	@ApiOperation({ title: 'Get company list' })
 	findAll(@Req() req, @Res() res) {
-		// this.companyService.findAll(req.user.TENANT_GUID).subscribe(
-		// 	data => { res.send(data); },
-		// 	err => { this.commonFunctionService.sendResErrorV3(err, res); }
-		// )
 		this.commonFunctionService.runGetServiceV2(this.companyService.findAll(req.user.TENANT_GUID), res);
 	}
 
@@ -51,14 +47,10 @@ export class CompanyController {
 	 */
 	@Get('/:id')
 	@ApiOperation({ title: 'Get one company' })
-	@ApiImplicitQuery({ name: 'id', description: 'Get by company guid', required: true })
+	@ApiImplicitParam({ name: 'id', description: 'Get by company guid', required: true })
 	findById(@Param('id') id, @Req() req, @Res() res) {
-		id = this.commonFunctionService.findIdParam(req, res, id);
 
-		if (id == null || id == '' || id == '{id}') {
-			// res.send('id not found') 
-			throw new NotFoundException('Id not found');
-		}
+		if (id == null || id == '' || id == '{id}') { throw new NotFoundException('Id not found'); }
 
 		this.companyService.findById(req.user.TENANT_GUID, id).subscribe(
 			data => { res.send(data); },
@@ -80,10 +72,9 @@ export class CompanyController {
 	@Post('/:name')
 	@ApiOperation({ title: 'Create company' })
 	createCompany(@Param('name') name: string, @Req() req, @Res() res) {
-		// this.companyService.create(req.user, name).subscribe(
-		// 	data => { if (data.status == 200) { res.send(data.data); } },
-		// 	err => { this.commonFunctionService.sendResErrorV2(res, 400, 'Fail to create resource'); }
-		// );
+
+		if (name == null || name == '' || name == '{name}') { throw new NotFoundException('Name not found'); }
+
 		this.commonFunctionService.runCreateService(this.companyService.create(req.user, name), res);
 	}
 
@@ -98,11 +89,6 @@ export class CompanyController {
 	@Patch()
 	@ApiOperation({ title: 'Update company' })
 	updateCompany(@Body() updateCompanyDTO: UpdateCompanyDTO, @Req() req, @Res() res) {
-		// this.companyService.update(req.user, updateCompanyDTO)
-		// 	.subscribe(
-		// 		data => { if (data.status == 200) { res.send(data.data); } },
-		// 		err => { this.commonFunctionService.sendResErrorV2(res, 400, 'Fail to update resource'); }
-		// 	)
 		this.commonFunctionService.runUpdateService(this.companyService.update(req.user, updateCompanyDTO), res);
 	}
 
@@ -116,9 +102,11 @@ export class CompanyController {
 	 */
 	@Delete('/:id')
 	@ApiOperation({ title: 'Delete company' })
-	@ApiImplicitQuery({ name: 'id', description: 'Delete by company guid', required: true })
+	@ApiImplicitParam({ name: 'id', description: 'Delete by company guid', required: true })
 	deleteRoleProfile(@Param('id') id, @Req() req, @Res() res) {
-		id = this.commonFunctionService.findIdParam(req, res, id);
+
+		if (id == null || id == '' || id == '{id}') { throw new NotFoundException('Id not found'); }
+
 		this.commonFunctionService.runUpdateService(this.companyService.deleteCompany(req.user, id), res);
 	}
